SP5: move gradient to top-level namespace
