+++ src/ScottPlot5/ScottPlot5/Primitives/GradientType.cs
@@ -1,0 +1,12 @@
+﻿namespace ScottPlot;
+
+/// <summary>
+/// Describes the geometry of a color gradient used to fill an area
+/// </summary>
+public enum GradiantType
+{
+    Linear,
+    Radial,
+    Sweep,
+    TwoPointConical
+}
